feat: move converters to one package
