+++ .gitignore
@@ -14,3 +14,5 @@ coverage.html
 .idea/*
 build/*
 data
+*.sqlite-shm
+*.sqlite-wal
